Resize patches to keep constant screen size.
The default world is 32x32 patches, with each patch being 13 pixels a
side. This means that the world is originally 416x416 pixels. So
dividing this number by image-size will give the appropriate patch size
to keep the world the same size on the screen.

Note: Documentation states that patch size is typically an integer, but
may be floating point.

We are also setting the patch size before resizing the world, to try to
keep any memory benefit it may give. (TODO)