Add logic to reduce amount of resizing
In general, we want to shrink then grow. So if the new world size is larger than the current one, shrinking the patch size then increasing the number of patches is the correct way.

Likewise, with a smaller new world size than current, changing the number of patches then growing the patch size will keep everything nice and small.