Successfully tested default memory limits
It appears that memory constrains are set by the world size and not the imported image size.

Having 2048 patches/side seems to exceed the memory limit of NetLogo, while 1024 is completely safe.

When loading any image, even the 5120 test, results in unchanging memory usage. Likely, this is due to NetLogo scaling the image to fit.

At 5120->1024 scaling, each patch represents 25 pixels square (5px/side). This *might* be okay for most applications, but that will be tested later.